# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "This is a Graduation Application Management System.  It is implemented on NodeJs with REST API Architecture. You can find\nout source code on\n[Github](https://github.com/microbraino/gradapp)\n",
        "version": "1.0.0",
        "title": "Graduate Application",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "abdulkadir.kahraman@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "description": "Heroku Server",
            "url": "http://commerchant.herokuapp.com"
        }
    ],
...
    "paths": {
        "/accounts/register": {
            "post": {
                "tags": [
                    "accounts"
                ],
                "summary": "Regist a new applicant to the system",
                "operationId": "registApplicant",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "requestBody": {
                    "$ref": "#/components/requestBodies/Register"
                }
            }
        },
        "/accounts/staffregister": {
            "post": {
                "tags": [
                    "accounts"
                ],
                "x-roles": [
                    "admin"
                ],
                "summary": "Regist a new staff to the addmisson database",
                "operationId": "registStaff",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StaffRegister"
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/accounts/password": {
            "patch": {
                "tags": [
                    "accounts"
                ],
                "x-roles": [
                    "admin",
                    "applicant",
                    "gradschool",
                    "department"
                ],
                "summary": "Update the password of current logged in account session",
                "operationId": "updatePassword",
                "requestBody": {
                    "$ref": "#/components/requestBodies/updatePassword"
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
    "components": {
...
        "requestBodies": {
            "Register": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "program": {
                                    "type": "string",
                                    "example": "5ecc3ecdf495810004264814"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "John"
                                },
                                "surname": {
                                    "type": "string",
                                    "example": "Doe"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "example@gmail.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "ab123"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": 15351234567
                                },
                                "address": {
                                    "type": "string",
                                    "example": "2757  Spring Haven Trail, Livingston, NJ"
                                }
                            }
                        }
                    }
                },
                "description": "Account object that needs to be added to the database",
                "required": true
            },
            "StaffRegister": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "John"
                                },
                                "surname": {
                                    "type": "string",
                                    "example": "Doe"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "example@gmail.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "ab123"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": 15351234567
                                },
                                "address": {
                                    "type": "string",
                                    "example": "2757  Spring Haven Trail, Livingston, NJ"
                                },
                                "role": {
                                    "type": "string",
                                    "enum": [
                                        "applicant",
                                        "admin",
                                        "gradschool",
                                        "department"
                                    ]
                                }
                            }
                        }
                    }
                },
                "required": true
            },
...
            "updatePassword": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "old": {
                                    "type": "string",
                                    "example": "ab123"
                                },
                                "new": {
                                    "type": "string",
                                    "example": "ab321"
                                }
                            }
                        }
                    }
                },
                "required": true
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}